csrf fixed

+++ app/views.py
@@ -1,4 +1,4 @@
-from flask import render_template, flash, redirect, url_for, request
+from flask import render_template, flash, redirect, url_for, request, session, abort
 from flask.ext.login import login_user, logout_user, current_user, login_required
 from app import app, db, lm
 from forms import EditForm, PostForm
@@ -6,6 +6,7 @@ from models import User, ROLE_USER, ROLE_ADMIN, Post
 from config import POSTS_PER_PAGE
 from datetime import datetime
 from oauth import OAuthSignIn
+import uuid
 
 @lm.user_loader
 def load_user(id):
@@ -36,12 +37,14 @@ def users(page = 1):
             title='List of users',
             users_list = users_list)
 
-@app.route('/delete/<int:post_id>')
+@app.route('/delete/<int:post_id>/<secret_id>')
 @login_required
-def delete_post(post_id):
+def delete_post(post_id, secret_id):
     if current_user.is_anonymous():
         return redirect(url_for('index'))
     post = Post.query.get(post_id)
+    if secret_id != session.get('secret_key', None):
+        abort(400)
     if post.author.id == current_user.id:
         db.session.delete(post)
         db.session.commit()
@@ -70,6 +73,7 @@ def oauth_callback(provider):
         db.session.commit()
         db.session.add(user.follow(user))
         db.session.commit()
+    session['secret_key'] = uuid.uuid4().hex
     login_user(user, True)
     return redirect(url_for('index'))
 
@@ -82,15 +86,18 @@ def login():
 
 @app.route('/logout')
 def logout():
+    session.pop('secret_key', None)
     logout_user()
     return redirect(url_for('index'))
 
-@app.route('/delete_account/<int:id>')
+@app.route('/delete_account/<int:id>/<secret_id>')
 @login_required
-def delete_account(id):
+def delete_account(id, secret_id):
     if not current_user.id == id:
         return redirect(url_for('index'))
     user = User.query.get(id)
+    if secret_id != session.get('secret_key', None):
+        abort(400)
     for post in user.posts:
         db.session.delete(post)
     db.session.delete(user)
@@ -130,9 +137,9 @@ def edit():
         return render_template('edit.html',
                 form = form)
 
-@app.route('/follow/<nickname>')
+@app.route('/follow/<nickname>/<secret_id>')
 @login_required
-def follow(nickname):
+def follow(nickname, secret_id):
     user = User.query.filter_by(nickname = nickname).first()
     if user == None:
         flash('User ' + nickname + ' not found')
@@ -140,6 +147,8 @@ def follow(nickname):
     if user == current_user:
         flash('You can\'t follow yourself!')
         return redirect(url_for('user', nickname = nickname))
+    if secret_id != session.get('secret_key', None):
+        abort(400)
     u = current_user.follow(user)
     if u is None:
         flash('Cannot follow ' + nickname)
@@ -149,9 +158,9 @@ def follow(nickname):
     flash('You are now following ' + nickname + '!')
     return redirect(url_for('user', nickname = nickname))
 
-@app.route('/unfollow/<nickname>')
+@app.route('/unfollow/<nickname>/<secret_id>')
 @login_required
-def unfollow(nickname):
+def unfollow(nickname, secret_id):
     user = User.query.filter_by(nickname = nickname).first()
     if user == None:
         flash('User ' + nickname + ' not found')
@@ -159,6 +168,8 @@ def unfollow(nickname):
     if user == current_user:
         flash('You can\'t unfollow yourself!')
         return redirect(url_for('user', nickname = nickname))
+    if secret_id != session.get('secret_key', None):
+        abort(400)
     u = current_user.unfollow(user)
     if u is None:
         flash('Cannot follow ' + nickname)
